requirements.txt
librosa==0.8.1
sounddevice==0.4.2
psycopg2-binary==2.9.1